Remove AppFormTwilogDate.vue
